fix(structs1): update color struct fields to use u8

+++ exercises/07_structs/structs1.rs
@@ -1,12 +1,12 @@
 struct ColorRegularStruct {
     // TODO: Add the fields that the test `regular_structs` expects.
     // What types should the fields have? What are the minimum and maximum values for RGB colors?
-    red: i16,
-    green: i16,
-    blue: i16,
+    red: u8,
+    green: u8,
+    blue: u8,
 }
 
-struct ColorTupleStruct(i16, i16, i16);
+struct ColorTupleStruct(u8, u8, u8);
 
 #[derive(Debug)]
 struct UnitStruct;
